last fixes
