--- src/components/modal/LoadingOverlay/index.js
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/no-unresolved
-import LoadingOverlay from './LoadingOverlay';
-
-export default LoadingOverlay;
